PSL_Test/requirements.txt
numpy
pyserial
opencv-python
PySide6
PyQt5
neurokit2
